Multiple Gsub

Performs multiple substitions in one string or multiple strings.

Usage

mgsub(pattern, replacement, x)

Arguments

pattern

character string containing a regular expression (or character string for fixed = TRUE) to be matched in the given character vector. Coerced by as.character to a character string if possible.

replacement

a replacement for matched pattern as in sub and gsub. See there for more information.

x

a character vector where matches are sought, or an object which can be coerced by as.character to a character vector. Long vectors are supported.

Value

a character vector of the same length and with the same attributes as x (after possible coercion to character).

See also

Examples

x <- c("ABC", "BCD", "CDE")
mgsub(pattern=c("B", "C"), replacement=c("X","Y"), x)
#> [1] "AXY" "XYD" "YDE"